requirements.txt
# Core ML framework - Latest TensorFlow with Python 3.12 support
tensorflow>=2.20.0

# AUCMEDI medical image classification framework - Latest version
aucmedi>=0.11.0

# Image processing and data handling
h5py>=3.11.0
pyvips>=3.0.0
opencv-python>=4.10.0
pillow>=11.0.0

# Data processing and utilities
numpy>=1.26.0,<3.0.0
pandas>=2.2.0
tqdm>=4.67.0

# Machine Learning libraries
scikit-learn>=1.5.0
scikit-image>=0.24.0
albumentations>=1.4.0

# Deep learning utilities (pinned by AUCMEDI)
keras-applications==1.0.8
classification-models-3D>=1.0.10
SimpleITK>=2.4.0
batchgenerators>=0.25
volumentations-aucmedi>=1.0.1

# Visualization and plotting
matplotlib>=3.9.0
plotnine==0.12.4

# Explainability and utilities
lime>=0.2.0.1
pooch>=1.8.0
pathos>=0.3.0

# Data download utilities
kaggle>=1.6.0

# Optional: Vision Transformer support (if needed, Python 3.12 not yet supported)
# vit-keras>=0.1.0  # Uncomment if using ViT models with Python 3.11 or lower
